refactor: application 테스트 도메인 패키지 하위로 이동
